Feat/refactor

+++ logigator-editor/src/app/services/simulation/worker-communication/worker-communication-wasm.service.ts
@@ -1,24 +1,33 @@
-import {Injectable, NgZone} from '@angular/core';
-import {Observable, Subject, timer} from 'rxjs';
-import {PowerChangesOutWire} from '../../../models/simulation/power-changes';
-import {ProjectsService} from '../../projects/projects.service';
-import {StateCompilerService} from '../state-compiler/state-compiler.service';
-import {WasmMethod, WasmRequest, WasmResponse} from '../../../models/simulation/wasm-interface';
-import {Element} from '../../../models/element';
-import {BoardState, BoardStatus, InputEvent} from '../../../models/simulation/board';
-import {takeWhile} from 'rxjs/operators';
-import {ErrorHandlingService} from '../../error-handling/error-handling.service';
-import {CompileError} from '../../../models/simulation/compile-error';
-import {ElementProviderService} from '../../element-provider/element-provider.service';
-import {AverageBuffer} from '../../../models/average-buffer';
-import {ElementTypeId} from '../../../models/element-types/element-type-ids';
-import {EastereggService} from '../../easteregg/easteregg.service';
-import {WorkerCommunicationServiceModel} from './worker-communication-service-model';
-import {InvalidPlugsError} from '../../../models/simulation/invalid-plugs-error';
+import { Injectable, NgZone } from '@angular/core';
+import { Observable, Subject, timer } from 'rxjs';
+import { PowerChangesOutWire } from '../../../models/simulation/power-changes';
+import { ProjectsService } from '../../projects/projects.service';
+import { StateCompilerService } from '../state-compiler/state-compiler.service';
+import {
+	WasmMethod,
+	WasmRequest,
+	WasmResponse
+} from '../../../models/simulation/wasm-interface';
+import { Element } from '../../../models/element';
+import {
+	BoardState,
+	BoardStatus,
+	InputEvent
+} from '../../../models/simulation/board';
+import { takeWhile } from 'rxjs/operators';
+import { ErrorHandlingService } from '../../error-handling/error-handling.service';
+import { CompileError } from '../../../models/simulation/compile-error';
+import { ElementProviderService } from '../../element-provider/element-provider.service';
+import { AverageBuffer } from '../../../models/average-buffer';
+import { ElementTypeId } from '../../../models/element-types/element-type-ids';
+import { EastereggService } from '../../easteregg/easteregg.service';
+import { WorkerCommunicationServiceModel } from './worker-communication-service-model';
+import { InvalidPlugsError } from '../../../models/simulation/invalid-plugs-error';
 
 @Injectable()
-export class WorkerCommunicationWasmService implements WorkerCommunicationServiceModel {
-
+export class WorkerCommunicationWasmService
+	implements WorkerCommunicationServiceModel
+{
 	private _powerSubjectsWires: Map<string, Subject<PowerChangesOutWire>>;
 	private _powerSubjectsWireEnds: Map<string, Subject<Map<Element, boolean[]>>>;
 	private _ioComponentsResetSubject: Map<string, Subject<void>>;
@@ -51,15 +60,17 @@ export class WorkerCommunicationWasmService implements WorkerCommunicationServic
 		private eastereggs: EastereggService
 	) {
 		this._powerSubjectsWires = new Map<string, Subject<PowerChangesOutWire>>();
-		this._powerSubjectsWireEnds = new Map<string, Subject<Map<Element, boolean[]>>>();
+		this._powerSubjectsWireEnds = new Map<
+			string,
+			Subject<Map<Element, boolean[]>>
+		>();
 		this._ioComponentsResetSubject = new Map<string, Subject<void>>();
 		this.initWorker();
 	}
 
-	private handleResponse(event: any): void {
+	private handleResponse(event: MessageEvent): void {
 		if (!this._initialized) {
-			if (event.data.initialized === undefined)
-				return;
+			if (event.data.initialized === undefined) return;
 
 			if (event.data.initialized === true) {
 				this._initialized = true;
@@ -76,7 +87,9 @@ export class WorkerCommunicationWasmService implements WorkerCommunicationServic
 
 			if (state.length !== this.stateCompiler.highestLinkId + 1) {
 				console.error(data);
-				this.errorHandling.showErrorMessage(`Response data length does not match component count`);
+				this.errorHandling.showErrorMessage(
+					`Response data length does not match component count`
+				);
 			}
 			this._dataCache = state;
 
@@ -84,12 +97,22 @@ export class WorkerCommunicationWasmService implements WorkerCommunicationServic
 				const powerChangesWire = new Map<string, PowerChangesOutWire>();
 				const powerChangesWireEnds = new Map<string, Map<Element, boolean[]>>();
 				for (const identifier of this._powerSubjectsWires.keys()) {
-					powerChangesWire.set(identifier, this.getWireState(identifier, state));
-					powerChangesWireEnds.set(identifier, this.getWireEndState(identifier, state));
+					powerChangesWire.set(
+						identifier,
+						this.getWireState(identifier, state)
+					);
+					powerChangesWireEnds.set(
+						identifier,
+						this.getWireEndState(identifier, state)
+					);
 				}
 				for (const projId of this._powerSubjectsWires.keys()) {
-					this._powerSubjectsWires.get(projId).next(powerChangesWire.get(projId));
-					this._powerSubjectsWireEnds.get(projId).next(powerChangesWireEnds.get(projId));
+					this._powerSubjectsWires
+						.get(projId)
+						.next(powerChangesWire.get(projId));
+					this._powerSubjectsWireEnds
+						.get(projId)
+						.next(powerChangesWireEnds.get(projId));
 				}
 			}
 
@@ -107,7 +130,9 @@ export class WorkerCommunicationWasmService implements WorkerCommunicationServic
 				this._worker.postMessage(request);
 			} else if (data.method === WasmMethod.run && this._mode === 'target') {
 				const timestamp = Date.now();
-				const ticks = ((timestamp - this._targetLastRun) * this._targetSpeed / 1000) + this._targetUnprocessedFraction;
+				const ticks =
+					((timestamp - this._targetLastRun) * this._targetSpeed) / 1000 +
+					this._targetUnprocessedFraction;
 				const ticksToCompute = Math.trunc(ticks);
 				const request: WasmRequest = {
 					method: WasmMethod.run,
@@ -127,20 +152,23 @@ export class WorkerCommunicationWasmService implements WorkerCommunicationServic
 			}
 
 			if (data.method === WasmMethod.status) {
-				this.ngZone.run(() => this._status = data.status);
+				this.ngZone.run(() => (this._status = data.status));
 			}
 		} else {
 			console.error('error', data);
 		}
 	}
 
-	public getWireState(identifier: string, data?: Uint8Array): Map<Element, boolean> {
-		if (!data)
-			data = this._dataCache;
+	public getWireState(
+		identifier: string,
+		data?: Uint8Array
+	): Map<Element, boolean> {
+		if (!data) data = this._dataCache;
 		const out = new Map<Element, boolean>();
-		if (!this.stateCompiler.wiresOnLinks.has(identifier))
-			return out;
-		for (const [link, wires] of this.stateCompiler.wiresOnLinks.get(identifier).entries()) {
+		if (!this.stateCompiler.wiresOnLinks.has(identifier)) return out;
+		for (const [link, wires] of this.stateCompiler.wiresOnLinks
+			.get(identifier)
+			.entries()) {
 			for (const wire of wires) {
 				out.set(wire, data[link] as unknown as boolean);
 			}
@@ -148,32 +176,36 @@ export class WorkerCommunicationWasmService implements WorkerCommunicationServic
 		return out;
 	}
 
-	public getWireEndState(identifier: string, data?: Uint8Array): Map<Element, boolean[]> {
-		if (!data)
-			data = this._dataCache;
+	public getWireEndState(
+		identifier: string,
+		data?: Uint8Array
+	): Map<Element, boolean[]> {
+		if (!data) data = this._dataCache;
 		const out = new Map<Element, boolean[]>();
-		if (!this.stateCompiler.wireEndsOnLinks.has(identifier))
-			return out;
-		for (const [link, wireEndOnComps] of this.stateCompiler.wireEndsOnLinks.get(identifier).entries()) {
+		if (!this.stateCompiler.wireEndsOnLinks.has(identifier)) return out;
+		for (const [link, wireEndOnComps] of this.stateCompiler.wireEndsOnLinks
+			.get(identifier)
+			.entries()) {
 			for (const wireEndOnComp of wireEndOnComps) {
 				const elem = wireEndOnComp.component;
 				if (!out.has(elem))
 					out.set(elem, new Array(elem.numInputs + elem.numOutputs));
-				out.get(elem)[wireEndOnComp.wireIndex] = data[link] as unknown as boolean;
-
+				out.get(elem)[wireEndOnComp.wireIndex] = data[
+					link
+				] as unknown as boolean;
 			}
 		}
 		return out;
 	}
 
 	public async init(): Promise<void> {
-		if (!this._initialized)
-			return;
+		if (!this._initialized) return;
 
 		const project = this.projectsService.mainProject;
 
 		try {
-			const compiledBoard = await this.stateCompiler.compileAsInt32Array(project);
+			const compiledBoard =
+				await this.stateCompiler.compileAsInt32Array(project);
 			if (!compiledBoard) {
 				throw new CompileError('ERROR.COMPILE.FAILED');
 			}
@@ -205,24 +237,34 @@ export class WorkerCommunicationWasmService implements WorkerCommunicationServic
 	}
 
 	private initWorker() {
-		if (this._worker)
-			this._worker.terminate();
+		if (this._worker) this._worker.terminate();
 
 		this._initialized = false;
-		this._worker = new Worker(new URL('../../../worker/simulation-worker/simulation.worker', import.meta.url), { type: 'module' });
+		this._worker = new Worker(
+			new URL(
+				'../../../worker/simulation-worker/simulation.worker',
+				import.meta.url
+			),
+			{ type: 'classic' }
+		);
 		this.ngZone.runOutsideAngular(() => {
-			this._worker.addEventListener('message', (event) => this.handleResponse(event as any));
+			this._worker.addEventListener('message', (event) =>
+				this.handleResponse(event)
+			);
 		});
 	}
 
 	private finalizeInit(compiledBoard: ArrayBuffer) {
-		this._worker.postMessage({
-			method: WasmMethod.init,
-			board: {
-				links: this.stateCompiler.highestLinkId + 1,
-				components: compiledBoard
-			}
-		} as WasmRequest, [ compiledBoard ]);
+		this._worker.postMessage(
+			{
+				method: WasmMethod.init,
+				board: {
+					links: this.stateCompiler.highestLinkId + 1,
+					components: compiledBoard
+				}
+			} as WasmRequest,
+			[compiledBoard]
+		);
 	}
 
 	public stop(): void {
@@ -237,8 +279,7 @@ export class WorkerCommunicationWasmService implements WorkerCommunicationServic
 	}
 
 	public start(): void {
-		if (this._mode === 'continuous')
-			return;
+		if (this._mode === 'continuous') return;
 
 		const request: WasmRequest = {
 			method: WasmMethod.run,
@@ -251,11 +292,9 @@ export class WorkerCommunicationWasmService implements WorkerCommunicationServic
 	}
 
 	public startTarget(target?: number): void {
-		if (this._mode === 'target')
-			return;
+		if (this._mode === 'target') return;
 
-		if (target && target >= 0)
-			this._targetSpeed = target;
+		if (target && target >= 0) this._targetSpeed = target;
 
 		const request: WasmRequest = {
 			method: WasmMethod.run,
@@ -269,8 +308,7 @@ export class WorkerCommunicationWasmService implements WorkerCommunicationServic
 	}
 
 	public startSync(): void {
-		if (this._mode === 'sync')
-			return;
+		if (this._mode === 'sync') return;
 
 		this._mode = 'sync';
 		this.registerStatusWatch();
@@ -278,21 +316,20 @@ export class WorkerCommunicationWasmService implements WorkerCommunicationServic
 	}
 
 	public setTarget(target: number) {
-		if (target >= 0)
-			this._targetSpeed = target;
+		if (target >= 0) this._targetSpeed = target;
 	}
 
 	private registerStatusWatch() {
 		const mode = this._mode;
 
 		this.ngZone.runOutsideAngular(() => {
-			timer(0, 1000).pipe(
-				takeWhile(() => mode === this._mode),
-			).subscribe(x => {
-				this._worker.postMessage({
-					method: WasmMethod.status
-				} as WasmRequest);
-			});
+			timer(0, 1000)
+				.pipe(takeWhile(() => mode === this._mode))
+				.subscribe(() => {
+					this._worker.postMessage({
+						method: WasmMethod.status
+					} as WasmRequest);
+				});
 		});
 	}
 
@@ -308,7 +345,11 @@ export class WorkerCommunicationWasmService implements WorkerCommunicationServic
 	}
 
 	public setFrameTime(frameTime: number): void {
-		this._frameAverage.push(frameTime > this._frameAverage.average + 100 ? this._frameAverage.average + 100 : frameTime);
+		this._frameAverage.push(
+			frameTime > this._frameAverage.average + 100
+				? this._frameAverage.average + 100
+				: frameTime
+		);
 
 		if (this._mode === 'sync') {
 			const request: WasmRequest = {
@@ -319,15 +360,21 @@ export class WorkerCommunicationWasmService implements WorkerCommunicationServic
 		}
 	}
 
-	public setUserInput(identifier: string, element: Element, state: boolean[]): void {
-		const index = this.stateCompiler.ioElemIndexes.get(identifier).get(element.id);
+	public setUserInput(
+		identifier: string,
+		element: Element,
+		state: boolean[]
+	): void {
+		const index = this.stateCompiler.ioElemIndexes
+			.get(identifier)
+			.get(element.id);
 		let inputEvent: InputEvent;
 		if (element.typeId === ElementTypeId.BUTTON) {
 			inputEvent = InputEvent.Pulse;
 		} else if (element.typeId === ElementTypeId.LEVER) {
 			inputEvent = InputEvent.Cont;
 		}
-		const stateBuffer = Int8Array.from(state as any).buffer;
+		const stateBuffer = Int8Array.from(state as unknown as number[]).buffer;
 
 		const request = {
 			method: WasmMethod.triggerInput,
@@ -337,14 +384,16 @@ export class WorkerCommunicationWasmService implements WorkerCommunicationServic
 				state: stateBuffer
 			}
 		} as WasmRequest;
-		this._worker.postMessage(request, [ stateBuffer ]);
+		this._worker.postMessage(request, [stateBuffer]);
 	}
 
 	public boardStateWires(projectId: string): Observable<PowerChangesOutWire> {
 		return this._powerSubjectsWires.get(projectId).asObservable();
 	}
 
-	public boardStateWireEnds(projectId: string): Observable<Map<Element, boolean[]>> {
+	public boardStateWireEnds(
+		projectId: string
+	): Observable<Map<Element, boolean[]>> {
 		return this._powerSubjectsWireEnds.get(projectId).asObservable();
 	}
 
@@ -362,8 +411,14 @@ export class WorkerCommunicationWasmService implements WorkerCommunicationServic
 
 	public subscribe(identifier: string): void {
 		if (!this._powerSubjectsWires.has(identifier)) {
-			this._powerSubjectsWires.set(identifier, new Subject<PowerChangesOutWire>());
-			this._powerSubjectsWireEnds.set(identifier, new Subject<Map<Element, boolean[]>>());
+			this._powerSubjectsWires.set(
+				identifier,
+				new Subject<PowerChangesOutWire>()
+			);
+			this._powerSubjectsWireEnds.set(
+				identifier,
+				new Subject<Map<Element, boolean[]>>()
+			);
 			this._ioComponentsResetSubject.set(identifier, new Subject<void>());
 		}
 	}
